Move 'nxpeaks' to 'nxrefine'
