+++ src/nxrefine/closest.py
