+++ src/nxrefine/plugins/refine/find_maximum.py
