+++ src/nxrefine/plugins/refine/find_peaks.py
