+++ src/nxrefine/plugins/refine/stack_images.py
