+++ src/nxrefine/scripts/nxmax.py
