+++ src/nxrefine/scripts/nxsetup.py
