+++ src/nxrefine/scripts/nxspec.py
